rename request specs and include routes as a comment
