+++ spec/requests/collections_spec.rb
@@ -1,0 +1,29 @@
+require "rails_helper"
+
+#      collections GET    /collections(.:format)                                                  collections#index
+#                  POST   /collections(.:format)                                                  collections#create
+#   new_collection GET    /collections/new(.:format)                                              collections#new
+#  edit_collection GET    /collections/:id/edit(.:format)                                         collections#edit
+#       collection GET    /collections/:id(.:format)                                              collections#show
+#                  PATCH  /collections/:id(.:format)                                              collections#update
+#                  PUT    /collections/:id(.:format)                                              collections#update
+#                  DELETE /collections/:id(.:format)                                              collections#destroy
+
+RSpec.describe "Collections" do
+  before do
+    sign_in create(:user)
+    build_list(:collection, 13) do |collection|
+      collection.save! # See https://dev.to/hernamvel/the-optimal-way-to-create-a-set-of-records-with-factorybot-createlist-factorybot-buildlist-1j64
+      create_list(:link, 1, linkable: collection)
+      create_list(:model, 1, collection: collection)
+    end
+  end
+
+  describe "GET /collections?page=2" do
+    it "returns paginated collections" do # rubocop:todo RSpec/MultipleExpectations
+      get "/collections?page=2"
+      expect(response).to have_http_status(:success)
+      expect(response.body).to match(/pagination/)
+    end
+  end
+end
