+++ spec/requests/devise/sessions_spec.rb
@@ -1,0 +1,8 @@
+require "rails_helper"
+
+#     new_user_session GET    /users/sign_in(.:format)                                                devise/sessions#new
+#         user_session POST   /users/sign_in(.:format)                                                devise/sessions#create
+# destroy_user_session DELETE /users/sign_out(.:format)                                               devise/sessions#destroy
+
+RSpec.describe "Devise::Sessions" do
+end
