+++ spec/requests/uploads_spec.rb
@@ -1,0 +1,7 @@
+require "rails_helper"
+
+#  uploads GET    /uploads(.:format)                                                      uploads#index
+#          POST   /uploads(.:format)                                                      uploads#create
+
+RSpec.describe "Uploads" do
+end
