+++ spec/requests/users/registrations_spec.rb
@@ -1,0 +1,12 @@
+require "rails_helper"
+
+# cancel_user_registration GET    /users/cancel(.:format)                                                 users/registrations#cancel
+#    new_user_registration GET    /users/sign_up(.:format)                                                users/registrations#new
+#   edit_user_registration GET    /users/edit(.:format)                                                   users/registrations#edit
+#        user_registration PATCH  /users(.:format)                                                        users/registrations#update
+#                          PUT    /users(.:format)                                                        users/registrations#update
+#                          DELETE /users(.:format)                                                        users/registrations#destroy
+#                          POST   /users(.:format)                                                        users/registrations#create
+
+RSpec.describe "Users::Registrations" do
+end
